SCRUM-1534 split view and logic
