+++ src/main/cliapp/src/components/GenericDataTable/useGenericDataTable.js
@@ -1,17 +1,10 @@
-import React, { useRef, useState, useEffect } from 'react';
+import { useRef, useState } from 'react';
 import { useQuery } from 'react-query';
-
-import { Column } from 'primereact/column';
-import { MultiSelect } from 'primereact/multiselect';
-
-import { FilterComponent } from './FilterComponent'
-import { EllipsisTableCell } from '../EllipsisTableCell';
-import { DataTableHeaderFooterTemplate } from "../DataTableHeaderFooterTemplate";
 
 import { SearchService } from '../../service/SearchService';
 import { useSessionStorage } from '../../service/useSessionStorage';
 
-import { trimWhitespace, returnSorted, filterColumns, orderColumns, reorderArray, setDefaultColumnOrder } from '../../utils/utils';
+import { trimWhitespace, returnSorted, reorderArray, setDefaultColumnOrder } from '../../utils/utils';
 import { useSetDefaultColumnOrder } from '../../utils/useSetDefaultColumnOrder';
 
 export const useGenericDataTable = ({ 
@@ -19,11 +12,9 @@ export const useGenericDataTable = ({
   tableName, 
   columns,  
   aggregationFields,
-  isEditable,
   curieFields,
   sortMapping,
   mutation,
-  isEnabled,
   setIsEnabled,
   toasts,
   initialColumnWidth,
@@ -245,77 +236,6 @@ export const useGenericDataTable = ({
   };
  
 
-  const createMultiselectComponent = (tableState,defaultColumnNames,isEnabled) => {
-      return (<MultiSelect
-          value={tableState.selectedColumnNames}
-          options={defaultColumnNames}
-          onChange={e => setSelectedColumnNames(e.value)}
-          style={{ width: '20em', textAlign: 'center' }}
-          disabled={!isEnabled}
-      />);
-  };
-
-  const header = (
-    <DataTableHeaderFooterTemplate
-        title = {tableName + " Table"}
-        tableState = {tableState}
-        defaultColumnNames = {defaultColumnNames}
-        multiselectComponent = {createMultiselectComponent(tableState,defaultColumnNames,isEnabled)}
-        onclickEvent = {(event) => resetTableState(event)}
-        isEnabled = {isEnabled}
-    />
-  );
-
-  const filterComponentTemplate = ({ type, filterName, fields, options, optionField, useKeywordFields, annotationsAggregations }) => {
-    return (
-      <FilterComponent
-        type={type}
-        filterName={filterName}
-        fields={fields}
-        isEnabled={isEnabled}
-        onFilter={onFilter}
-        options={options}
-        optionField={optionField}
-        useKeywordFields = {useKeywordFields}
-        aggregationFields={aggregationFields}
-        tableState={tableState}
-        annotationsAggregations={annotationsAggregations}
-        endpoint={endpoint}
-      />
-    );
-  };
-
-  const obsoleteTemplate = (rowData) => {
-    if (rowData && rowData.obsolete !== null && rowData.obsolete !== undefined) {
-      return <EllipsisTableCell>{JSON.stringify(rowData.obsolete)}</EllipsisTableCell>
-    }
-  };
-
-  useEffect(() => {
-    const filteredColumns = filterColumns(columns, tableState.selectedColumnNames);
-    const orderedColumns = orderColumns(filteredColumns, tableState.selectedColumnNames);
-    setColumnList(
-      orderedColumns.map((col) => {
-        return <Column
-          style={{'minWidth':`${columnWidths[col.field]}vw`, 'maxWidth': `${columnWidths[col.field]}vw`}}
-          headerClassName='surface-0'
-          columnKey={col.field}
-          key={col.field}
-          field={col.field}
-          header={col.header}
-          body={col.field === "obsolete" ? obsoleteTemplate : col.body}
-          sortable={isEnabled}
-          filter
-          editor={col.editor}
-          showFilterMenu={false}
-          filterElement={() => filterComponentTemplate({...col.filterElement})}
-        />;
-      })
-    );
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [tableState, isEnabled, columnWidths]);
-
-
   const resetTableState = () => {
     let _tableState = {
       ...initialTableState,
@@ -352,44 +272,26 @@ export const useGenericDataTable = ({
     setColumnWidths(_columnWidths);
   };
 
-  const tableProps = { 
-          dataKey:'id',
-          value:entity,
-          header:header, 
-          ref:dataTable, 
-          filterDisplay:"row", 
-          scrollHeight:"62vh", 
-          scrollable: true,
-          tableClassName:'p-datatable-md',
-          editMode: "row",
-          onRowEditInit: onRowEditInit,
-          onRowEditCancel: onRowEditCancel, 
-          onRowEditSave: onRowEditSave,
-          editingRows: editingRows, 
-          onRowEditChange:onRowEditChange,
-          sortMode:"multiple", 
-          removableSort:true, 
-          onSort:onSort, 
-          multiSortMeta:tableState.multiSortMeta,
-          onColReorder:colReorderHandler, 
-          reorderableColumns: true, 
-          resizableColumns: true, 
-          columnResizeMode:"expand", 
-          showGridlines: true, 
-          onColumnResizeEnd:handleColumnResizeEnd,
-          paginator: true, 
-          totalRecords:totalRecords, 
-          onPage:onLazyLoad, 
-          lazy: true, 
-          first:tableState.first,
-          paginatorTemplate:"CurrentPageReport FirstPageLink PrevPageLink PageLinks NextPageLink LastPageLink RowsPerPageDropdown",
-          currentPageReportTemplate:"Showing {first} to {last} of {totalRecords}", 
-          rows:tableState.rows, 
-          rowsPerPageOptions:[10, 20, 50, 100, 250, 1000],
-  }
-
   return {
-    tableProps,
-    columnList
+    setSelectedColumnNames, 
+    defaultColumnNames, 
+    tableState, 
+    resetTableState, 
+    onFilter, 
+    setColumnList, 
+    columnWidths, 
+    entity, 
+    dataTable, 
+    editingRows, 
+    onRowEditInit, 
+    onRowEditCancel,
+    onRowEditSave,
+    onRowEditChange,
+    onSort,
+    colReorderHandler,
+    handleColumnResizeEnd,
+    totalRecords,
+    onLazyLoad,
+    columnList,
   };
 };
